Apply most patches provided to me over the years -- including a patch to support compilation on 64-bit architectures and some fixes for FreeBSD and Mac OS X.
git-svn-id: http://svn.code.sf.net/p/dmucs/code/trunk@20 4fab6e03-1dcd-4e32-a435-d2e2fcd52ff7

+++ dmucs/dmucs_msg.cc
@@ -130,7 +130,7 @@ DmucsHostReqMsg::handle(Socket *sock, const char *buf)
 
 	fprintf(stderr, "Giving out %s\n", resolved_name.c_str());
 
-	db->assignCpuToClient(cpuIpAddr, dprop_, (unsigned int) sock);
+	db->assignCpuToClient(cpuIpAddr, dprop_, sock);
 #if 0
 	fprintf(stderr, "The databases are now:\n");
 	db->dump();
